added getters and setters for wb and backlight

+++ eyes_on_me/backlight.py
@@ -7,4 +7,4 @@ def set_backlight(level):
 
 
 def get_backlight():
-    return int(string.strip(subprocess.check_output("xbacklight", shell=True)))
+    return float(string.strip(subprocess.check_output("xbacklight", shell=True)))
